+++ src/common/communication.js
@@ -1,7 +1,7 @@
 goog.module('omid.common.Communication');
 
-const {CommunicationType} = goog.require('omid.common.constants');
 const InternalMessage = goog.require('omid.common.InternalMessage');
+const {CommunicationType} = goog.require('omid.common.constants');
 
 /**
  * A send and forget style communication between two generic instances.